+++ Models/OrderItem.cs
@@ -7,7 +7,7 @@
         public virtual Order Order { get; set; }
         public int ProductId { get; set; }
         public virtual Product Product { get; set; }
-        public decimal Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
+        public int Quantity { get; set; }
+        public int UnitPrice { get; set; }
     }
 }
